refactor action creator pattern
